% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksApp-usage.R
\name{calculate_catch_record_usage}
\alias{calculate_catch_record_usage}
\title{Calculate catch record usage by week}
\usage{
calculate_catch_record_usage(
  trips,
  users,
  countries = c("Kenya", "Zanzibar", "Mozambique")
)
}
\arguments{
\item{trips}{tibble of trip data}

\item{users}{tibble of user data}

\item{countries}{character vector of countries to include}
}
\value{
tibble of catch record usage by week
}
\description{
Calculate catch record usage by week
}
