% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksApp-usage.R
\name{create_usage_heatmap}
\alias{create_usage_heatmap}
\title{Create usage heatmap plot}
\usage{
create_usage_heatmap(catch_record_usage, top_n = 50)
}
\arguments{
\item{catch_record_usage}{tibble of catch record usage data}

\item{top_n}{integer, number of top boats to display (default 50)}
}
\value{
ggplot object
}
\description{
Create usage heatmap plot
}
