% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksApp-usage.R
\name{create_usage_table}
\alias{create_usage_table}
\title{Create interactive reactable}
\usage{
create_usage_table(usage_summary)
}
\arguments{
\item{usage_summary}{tibble of usage summary data}
}
\value{
reactable object
}
\description{
Create interactive reactable
}
