% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lbb-engine.R
\name{.lbb_bb0}
\alias{.lbb_bb0}
\title{Relative biomass B/B0 = fished / unfished per-recruit biomass (weight prop L^b)}
\usage{
.lbb_bb0(lmid, linf, lc, alpha, mk, fk, b = 3)
}
\description{
Relative biomass B/B0 = fished / unfished per-recruit biomass (weight prop L^b)
}
\keyword{internal}
