% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lbb-engine.R
\name{.lbb_numbers}
\alias{.lbb_numbers}
\title{Relative numbers-at-length under the LBB equilibrium model}
\usage{
.lbb_numbers(lmid, linf, lc, alpha, mk, fk)
}
\description{
Relative numbers-at-length under the LBB equilibrium model
}
\keyword{internal}
