% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{download_cloud_file}
\alias{download_cloud_file}
\title{Download File from Cloud Storage}
\usage{
download_cloud_file(name, provider, options, file = name)
}
\arguments{
\item{name}{A character vector of object names in cloud storage.}

\item{provider}{A character string specifying the cloud storage provider.}

\item{options}{A named list of cloud storage provider options.}

\item{file}{A character vector of local file paths. Defaults to object names.}
}
\value{
A character vector of local file paths.
}
\description{
Downloads one or more files from cloud storage.
}
\keyword{storage}
