% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable.R
\name{fetch_asset}
\alias{fetch_asset}
\title{Fetch and Filter Asset Data from Airtable}
\usage{
fetch_asset(table_name = NULL, select_cols = NULL, conf = NULL)
}
\arguments{
\item{table_name}{Character. Name of the Airtable table to fetch.}

\item{select_cols}{Character vector. Column names to select from the table.}

\item{conf}{Configuration object from read_config().}
}
\value{
A filtered and selected data frame from Airtable
}
\description{
Retrieves data from a specified Airtable table
}
\keyword{helper}
\keyword{preprocessing}
