% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{mdb_collection_pull}
\alias{mdb_collection_pull}
\title{Retrieve Data from MongoDB}
\usage{
mdb_collection_pull(
  connection_string = NULL,
  collection_name = NULL,
  db_name = NULL
)
}
\arguments{
\item{connection_string}{A character string specifying the MongoDB connection URL.}

\item{collection_name}{A character string specifying the name of the collection.}

\item{db_name}{A character string specifying the name of the database.}
}
\value{
A data frame containing all documents from the specified collection.
}
\description{
Connects to a MongoDB database and retrieves all documents from a specified
collection, maintaining the original column order if a metadata document
is present.
}
\keyword{storage}
