% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{pelagic_refresh_token}
\alias{pelagic_refresh_token}
\title{Refresh Authentication Token}
\usage{
pelagic_refresh_token(
  refresh_token,
  base_url = "https://analytics.pelagicdata.com"
)
}
\arguments{
\item{refresh_token}{Character. Refresh token obtained from pelagic_auth()}

\item{base_url}{Character. Base URL for the API. Default is "https://analytics.pelagicdata.com"}
}
\value{
List containing the new access token and refresh token
}
\description{
Refreshes an expired access token using the refresh token.
}
