% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracks.R
\name{plot_fishing_effort}
\alias{plot_fishing_effort}
\title{Create an interactive Leaflet map of fishing effort on H3 grid}
\usage{
plot_fishing_effort(
  grid_data,
  metric = c("fishing_pings", "unique_trips"),
  title = "Fishing Effort"
)
}
\arguments{
\item{grid_data}{An sf object with h3_index, fishing_pings, unique_trips columns}

\item{metric}{Column to visualize: "fishing_pings" or "unique_trips"}

\item{title}{Map title}
}
\value{
A leaflet htmlwidget
}
\description{
Create an interactive Leaflet map of fishing effort on H3 grid
}
