% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tracks.R
\name{plot_fishing_grounds}
\alias{plot_fishing_grounds}
\title{Create an interactive Leaflet map of fishing grounds}
\usage{
plot_fishing_grounds(fishing_grounds, title = "Fishing Grounds")
}
\arguments{
\item{fishing_grounds}{An sf POLYGON object with ground_id and area_km2 columns}

\item{title}{Map title}
}
\value{
A leaflet htmlwidget
}
\description{
Create an interactive Leaflet map of fishing grounds
}
