% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksApp-usage.R
\name{prepare_usage_summary}
\alias{prepare_usage_summary}
\title{Prepare usage summary table}
\usage{
prepare_usage_summary(catch_record_usage)
}
\arguments{
\item{catch_record_usage}{tibble of catch record usage data}
}
\value{
tibble of usage summary statistics
}
\description{
Prepare usage summary table
}
