% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_pds_tracks}
\alias{preprocess_pds_tracks}
\title{Preprocess Pelagic Data Systems (PDS) Track Data}
\usage{
preprocess_pds_tracks(
  log_threshold = logger::DEBUG,
  grid_size = 500,
  package = "coasts"
)
}
\arguments{
\item{log_threshold}{The logging threshold to use. Default is logger::DEBUG.}

\item{grid_size}{Numeric. Size of grid cells in meters (100, 250, 500, or 1000).}

\item{package}{Name of the package whose \code{inst/conf.yml} to read. Defaults
to \code{"coasts"}.}
}
\value{
None (invisible).
}
\description{
Downloads raw GPS tracks and creates a gridded summary of fishing activity.
}
\keyword{preprocessing}
\keyword{workflow}
