% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksApp-usage.R
\name{pull_catch_events}
\alias{pull_catch_events}
\title{Pull catch events data from MongoDB}
\usage{
pull_catch_events(conf)
}
\arguments{
\item{conf}{Configuration object with MongoDB connection details}
}
\value{
tibble of catch events
}
\description{
Pull catch events data from MongoDB
}
