% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksApp-usage.R
\name{pull_users_data}
\alias{pull_users_data}
\title{Pull users data from MongoDB}
\usage{
pull_users_data(conf)
}
\arguments{
\item{conf}{Configuration object with MongoDB connection details}
}
\value{
tibble of users
}
\description{
Pull users data from MongoDB
}
