% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{upload_cloud_file}
\alias{upload_cloud_file}
\title{Upload File to Cloud Storage}
\usage{
upload_cloud_file(file, provider, options, name = file)
}
\arguments{
\item{file}{A character vector of file paths to upload.}

\item{provider}{A character string specifying the cloud storage provider.}

\item{options}{A named list of cloud storage provider options.}

\item{name}{A character vector of names to assign files in cloud storage.
Defaults to local filenames.}
}
\value{
A list of upload metadata.
}
\description{
Uploads one or more local files to a cloud storage bucket.
}
\keyword{storage}
