% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao-estimation.R
\name{flag_quality}
\alias{flag_quality}
\title{Flag estimates that fail the FAO 15 percent relative-error target}
\usage{
flag_quality(estimates, threshold = 0.15, warn_at = 0.2)
}
\arguments{
\item{estimates}{Tibble with `re_cpue` and/or `re_total_catch`.}

\item{threshold}{RE above which we fail. Default 0.15 (FAO target).}

\item{warn_at}{RE above which we warn. Default 0.20.}
}
\value{
The input tibble with quality columns appended.
}
\description{
Adds three columns:
  `quality_cpue`    - pass / warn / fail based on `re_cpue`.
  `quality_catch`   - same, based on `re_total_catch`.
  `quality_overall` - worst of the two.
}
