% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fishery.R
\name{process_trip_data}
\alias{process_trip_data}
\title{Process Trip Data with District Information}
\usage{
process_trip_data(conf = NULL, devices_table = NULL, imei_list = NULL)
}
\arguments{
\item{conf}{Configuration parameters containing:
- pds$token: API token for PDS access
- pds$secret: API secret for PDS access}

\item{devices_table}{Data frame containing device metadata with at minimum
an `imei` column. Used to join district information to trips.}

\item{imei_list}{Vector of IMEI numbers for devices to retrieve trip data for}
}
\value{
A processed data frame with:
  - district: District name (from community metadata)
  - community: Community name
  - trip: Trip identifier
  - boat: Boat identifier
  - landing_date: Date of trip landing
  - date_month: First day of landing month
  - duration_hrs: Trip duration in hours
}
\description{
Retrieves and processes trip data from the PDS API by adding district information
from community metadata. Filters out unrealistic trips longer than 48 hours to
remove outliers and prepares data for monthly statistics calculation.
}
\details{
The function:
- Calls `get_trips()` API to retrieve trip data for specified IMEIs
- Converts timestamps to landing dates and monthly periods
- Calculates trip duration in hours from seconds
- Joins with community metadata to add district information
- Filters out unrealistic trips (>48 hours)
- Standardizes column names and structure
}
\examples{
\dontrun{
# Process trip data with parameters and device list
conf <- read_config()
trips_stats <- process_trip_data(
  conf = conf,
  imei_list = imei_list
)

# Check data structure
str(trips_stats)
}

}
\seealso{
* [calculate_monthly_trip_stats()] for using processed trip data
* [get_trips()] for the underlying API call
* [get_metadata()] for community metadata retrieval
}
\keyword{preprocessing}
\keyword{workflow}
