% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_config}
\alias{read_config}
\title{Read configuration file}
\usage{
read_config(package = "coasts")
}
\arguments{
\item{package}{Name of the package whose `inst/conf.yml` to read. Defaults
to `"coasts"`. Downstream packages that ship their own `conf.yml` should
pass their own package name here so that their configuration is loaded
instead of the `coasts` defaults.}
}
\value{
the environment parameters
}
\description{
Reads configuration file in `conf.yml` and adds some logging lines. Wrapped
for convenience
}
\keyword{helper}
